+++ djdt_api_requests/panels/api_requests.py
@@ -2,6 +2,8 @@
 from __future__ import unicode_literals
 
 from debug_toolbar.panels import Panel
+from django.utils.translation import ugettext_lazy as _, ungettext
+
 import requests.sessions
 
 
@@ -9,7 +11,12 @@ def patch_send(send_method, panel):
 
     def patched_send(self, request, **kwargs):
         response = send_method(self, request, **kwargs)
-        panel.record(elapsed=response.elapsed.total_seconds())
+        panel.record(
+            method=request.method,
+            url=request.url,
+            status_code=response.status_code,
+            elapsed=(1000 * response.elapsed.total_seconds())
+        )
         return response
     return patched_send
 
@@ -17,26 +24,25 @@ def patch_send(send_method, panel):
 class ApiRequestsPanel(Panel):
     nav_title = 'API Requests'
 
-    # Content would be good to add in a future version
-    has_content = False
-
     title = 'API Requests'
+
+    template = 'djdt_api_requests/panels/api_requests.html'
 
     def __init__(self, *args, **kwargs):
         super(ApiRequestsPanel, self).__init__(*args, **kwargs)
         self._saved_send = None
-        self._num_requests = 0
-        self._times = []
+        self._requests = []
 
-    def record(self, elapsed):
-        # We will probably want to add URLs to this in the future
-        self._num_requests += 1
-        self._times.append(elapsed)
+    def record(self, **kwargs):
+        self._requests.append(kwargs)
 
     @property
     def nav_subtitle(self):
-        total_time = sum(self._times)
-        return "%d API requests in %.2fms" % (self._num_requests, 1000 * total_time)
+        total_time = sum(r['elapsed'] for r in self._requests)
+        num_requests = len(self._requests)
+        return ungettext("%d API request in %.2fms",
+                         "%d API requests in %.2fms", num_requests,
+                         ) % (len(self._requests), total_time)
 
     def enable_instrumentation(self):
         # monkeypatch the requests library
@@ -52,7 +58,8 @@ class ApiRequestsPanel(Panel):
 
     def generate_stats(self, request, response):
         self.record_stats({
-            'num_requests': self._num_requests,
-            'total_time': sum(self._times)
+            'num_requests': len(self._requests),
+            'total_time': sum(r['elapsed'] for r in self._requests),
+            'requests': self._requests
         })