+++ CHANGELOG.md
@@ -6,4 +6,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [0.1.0] - 2018-06-11
 ### Added
-- Total request count and call time taken to smmary display in panel
+- Total request count and call time taken to summary display in panel
+- Add detail view